Move looselyEscapeAsStringLiteralForJs to Utils.kt so it can be reused through the code
