+++ rich-html-editor/src/main/java/com/infomaniak/lib/richhtmleditor/Utils.kt
@@ -1,0 +1,37 @@
+/*
+ * Infomaniak Rich HTML Editor - Android
+ * Copyright (C) 2024 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.lib.richhtmleditor
+
+// TODO: This method might not be enough to escape user inputs and prevent access to JS code execution
+fun looselyEscapeAsStringLiteralForJs(string: String): String {
+    val stringBuilder = StringBuilder("`")
+
+    string.forEach {
+        val char = when (it) {
+            '`' -> "\\`"
+            '\\' -> "\\\\"
+            '$' -> "\\$"
+            else -> it
+        }
+        stringBuilder.append(char)
+    }
+
+    stringBuilder.append("`")
+
+    return stringBuilder.toString()
+}
